+++ __tests__/components.js
@@ -11,37 +11,43 @@ import icon from './components/button/features/icon';
 import clickValue from './components/button/features/click-value';
 import highlightFlags from './components/features/highlite-flags';
 
-describe('components tests', () => {
-  it('forgeed properties', () => {
+describe('Components tests', () => {
+  it('Should render component correctly', () => {
     const features = forge(icon, clickValue, highlightFlags);
     const FeaturesButton = features(Button, 'FeaturesButton');
 
-    const initialProps = {
+    const componentProps = {
       alert: true,
       clickValue: 'yo',
       iconPosition: 'right',
       icon: 'mail',
     };
 
-    const component = renderer.create(<FeaturesButton {...initialProps} />);
+    const component = renderer.create(<FeaturesButton {...componentProps} />);
     const tree = component.toJSON();
 
     const propsKeys = Object.keys(tree.props);
 
-    // used toContain expect toEqual bacause of componen's default props
+    // Expect result props at already rendered component
     expect(propsKeys).toEqual([
+      // There no Button props at componentProps so all default props are apllied
       ...Object.keys(Button.defaultProps),
+      // from clikcValue feature
       'onClick',
+      // from highlightFlags feature
       'style',
     ]);
 
 
     expect(tree.props.style).toEqual({
+      /**
+       * highlightFlags provide style color:red if "alert" flag exists
+       */
       color: 'red',
     });
   });
 
-  it('simulate click with clickValue feature', () => {
+  it('Should wrap onClick wiht hof (clikcValue feature)', () => {
     const onClickMock = jest.fn();
 
     const features = forge(clickValue, highlightFlags);