+++ __tests__/components/button/index.js
@@ -1,5 +1,7 @@
 import React, { PropTypes } from 'react';
-
+import { ThemeProp } from '../../../lib';
+// import { base, style } from './button.css';
+//
 const Button = ({
   children,
   ...props
@@ -8,6 +10,10 @@ const Button = ({
 Button.propTypes = {
   children: PropTypes.node,
   disabled: PropTypes.bool,
+  theme: ThemeProp({
+    base: PropTypes.string,
+    style: PropTypes.string,
+  }),
 };
 
 Button.defaultProps = {